+++ apps/backend/backend_startup.c
@@ -86,7 +86,7 @@ db_merge(clicon_handle h,
     cxobj *xt = NULL;
     
     /* Get data as xml from db1 */
-    if (xmldb_get0(h, (char*)db1, NULL, NULL, 0, &xt, NULL) < 0)
+    if (xmldb_get0(h, (char*)db1, YB_MODULE, NULL, NULL, 0, &xt, NULL) < 0)
 	goto done;
     /* Merge xml into db2. Without commit */
     retval = xmldb_put(h, (char*)db2, OP_MERGE, xt, clicon_username_get(h), cbret);