+++ lib/clixon/clixon_datastore.h
@@ -51,7 +51,7 @@ int xmldb_connect(clicon_handle h);
 int xmldb_disconnect(clicon_handle h);
  /* in clixon_datastore_read.[ch] */
 int xmldb_get(clicon_handle h, const char *db, cvec *nsc, char *xpath, cxobj **xtop);
-int xmldb_get0(clicon_handle h, const char *db,
+int xmldb_get0(clicon_handle h, const char *db, yang_bind yb,
 	       cvec *nsc, const char *xpath,
 	       int copy, cxobj **xtop, modstate_diff_t *msd); 
 int xmldb_get0_clear(clicon_handle h, cxobj *x);